% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kre.R
\name{kernelRatioEstimation}
\alias{kernelRatioEstimation}
\title{Kernel Ratio estimation: adaptive bandwidth kernel density estimation}
\usage{
kernelRatioEstimation(
  pop_grid,
  point,
  max_bw,
  min_bw,
  min_pop,
  cell_size,
  kernel = c("gaussian", "uniform", "quartic", "triweight", "epanechnikov")
)
}
\arguments{
\item{pop_grid}{A \code{\link[raster]{Raster-class}} contains population info}

\item{max_bw}{maximum searching radius (unit in mile)}

\item{min_bw}{minimum searching radius (unit in mile)}

\item{min_pop}{minimum population search radius needs to cover for each point}

\item{cell_size}{resolution of output raster file (unit in meter)}
}
\value{
\code{\link[raster]{Raster-class}}
}
\description{
Kernel Ratio estimation: adaptive bandwidth kernel density estimation
}
